Cubic VS westwood test added
Either test is easily changable for other TCPs, and this gives a
consistent way to observe both the effects of QoS marking and
the TCP in use...

Longer term, perhaps being able to override the basic TCP in use
would be good for some tests.

-T TCP_NAME,TCP_NAME

Would need to capture the TCP in the json then.

A problem I hit with this was that:

/proc/sys/net/ipv4/tcp_allowed_congestion_control

Did not allow westwood on the other side I had up and thus I got back
results that didn't look like westwood with no other indication
that the test was messed up. Perhaps netperf/netserver should fail
or return an error message when it can't set the requested TCP.