a torrent-like emulation of the worse phase of torrent
6 up, 6 down, (the normal number of flows on a torrent) using the
ledbat congestion control module, compared against 1 up, 1 down, and
the usual slew of measurement flows. A possibly more accurate test
would be to rotate 1 connection every 15 seconds, as torrent does,
and to use wildly varying RTTs.

This is a simulation of the download/upload phase of a torrent
connection, with the not-quite-standard CS1 markings. It would be
better to compare it against http accesses than another steady state
set of flows, but it may be interesting in drop tail or pie scenarios.

Amusingly, the tcp-ledbat kernel code broke entirely on linux 3.13 and
later, and I haven't fixed it yet.

I have had some hope that with a reduced target delay (say 10ms),
torrent could be made to scavange again in this environment.