tests: tcp_upload_prio - for testing high priority queue in qdiscs.
The added test tcp_upload_prio, is designed for testing the Linux
Kernels current default qdisc pfifo_fast and qdisc prio.

These qdiscs have a high priority queue, which uses the ToS bits
for prioritation.  Thus, the test sends ICMP ping packet with
these ToS bits, while two TCP uploads are running.

The focus of the test is to measure the effectiveness of the high
prio queue, as this demonstrate the Head-of-Line blocking
occuring due to 1) packets in the HW TX ring buffer, and 2) in
the qdisc layers requeue mechanism.

Special care need to be take for using this on the default
qdisc MQ which have pfifo_fast assigned to every HW queue.

Setup requirements
 1. IRQ align CPUs to NIC HW queues
  E.g. via: set_irq_affinity ethX
  From: https://github.com/netoptimizer/network-testing/

 2. Force netperf-wrapper subcommands to run the same CPU
  E.g: taskset -c 2 ./netperf-wrapper -H IP tcp_upload_prio

This will force all measurements to go through the same qdisc.  This
is needed so the ping/latency tests measures the real property of
the qdisc and Head-of-Line blocking effect.

Signed-off-by: Jesper Dangaard Brouer <brouer@redhat.com>