Support chaining of runners in test specifications.
This adds support for chaining runners; meaning that one runner can be
set to start when another finishes. This is done using the 'run_after'
keyword in the specification of a runner in the DATA_SETS mapping. The
contents of 'run_after' is the name of another runner; the runner in
which the 'run_after' keyword is specified will wait for this other
runner to exit before starting.

The 'run_after' keyword can be combined with the 'delay' keyword, in
which case the delay is applied once the other runner has exited.
Multiple runners can wait for the same runner to exit, and no checks are
done to avoid loops or unreachable states; it is up to the test
specification to make sure all runners actually execute eventually.