Support 'watchdog' runners which kill other runners on exit.
This adds support for a 'watchdog' feature in the runners, where a
runner can specify the 'kill_after' keyword with the name of another
runner. This has the effect that if runner A specifies runner B in the
kill_after keyword, runner A will be forcefully killed (by SIGTERM) once
runner B exits.

A new 'timer' type of runner can be used along with this feature to
specify a simple timer-based watchdog feature. The timer runner takes a
timeout as parameter, and will simply emit its exit signal after that
amount of seconds has passed. The kill_after keyword can specify
arbitrary runners, however, so it can also be used to, e.g., implement a
bandwidth monitor (by having a program monitor consumed bandwidth and
simply exit if a threshold is passed, thereby triggering the killing of
other runners).