Default the GUI to save plots in the current dir.
This tries to detect if matplotlib is configured to use the default
value for the `savefig.directory` rc value. If it is, it is overridden
to the cwd of the GUI when it starts.

Additionally, the first time a file is loaded through the GUI, the
savefig dir will be set to that file's directory. This catches the case
where the GUI is started in the home dir and subsequently loads a file
from another dir.

Closes #12.