Bring Iperf support up-to-date.
Updates Iperf support to use a find_iperf method that will detect
whether the version of Iperf is new enough to support sub-second
timestamping, and complain if not. The Iperf parser now assumes this is
the case and doesn't rely on patching Iperf anymore. It also produces
metadata and raw values correctly.

No tests actually use Iperf presently, but adding it in is now
straight-forward.