Fix runners module import on Windows.
The function definitions in the runners module relied on parts of the os
module that are not available on Windows. This is now conditioned on the
platform, so running Flent should at least work on Windows. However,
running tests requires some reworking of the process management code; so
for now we give an explicit error when trying to run a test on Windows.