Add tcp_nup and derived tests.
This adds a new TCP upload test where the number of streams can be
specified as the test parameter 'upload_streams'. In addition, derived
tests for 2,4,6,8 and 12 streams are added which simply sets the test
parameter and then includes the nup test.

Thanks to cklatsky <carl_klatsky@cable.comcast.com> for the original
work of implementing the tcp_*up.conf tests.