gui: Defer canvas resize event until tab is shown, add async timer.
This adds an async timer to track background drawing while the tab is
active. The async callback can be missed sometimes; the timer makes sure
we update the display anyway.

This also gets rid of the error callback, which restores Python 2
compatibility and prevents a torrent of popups if something goes wrong
on initial load.