runners: Suppress duplicate SsRunner instances
When running multiple flows to the same host, ss_iterate will output socket
statistics data for all of them. This means that having multiple instances of
the script run in this case is redundant, and can hurt performance
significantly.

So instead, detect when duplicate instances of SsRunner are started (i.e.
multiple instances with the same arguments), and reuse the output of the first
one instead of spawning duplicates.

Signed-off-by: Toke Høiland-Jørgensen <toke@toke.dk>