tests: Express old rtt_fair tests as special cases of rtt_fair_var
This consolidates the test definitions for the old rtt_fair* tests to use the
same definition as the rtt_fair_var tests, just configured statically with the
same number of hosts as before. This makes the test parameters (such as
control_hosts) available to these old tests as well, which fixes #114.

This also adds a new test parameter, 'markings', which is a list of diffserv
markings to set for each flow, which can be turned into a fix for #65.

As a nice side effect, this removes about a thousand lines of duplicate test
definition code.

Signed-off-by: Toke Høiland-Jørgensen <toke@toke.dk>