runners: Add irtt runner for UDP RTT measurements
This is a new UDP RTT runner that uses https://github.com/peteheist/irtt to
measure UDP roundtrip times.

This also adds a new rrul_be_irtt test that modifies the rrul_be test to use
irtt instead of netperf for the UDP measurement flows. This is mostly a stopgap
measure for testing until a better way of choosing between netperf and irtt for
UDP RTT tests is implemented.

Signed-off-by: Toke Høiland-Jørgensen <toke@toke.dk>