runners: Add DelegatingRunner for automatically selecting UDP RR binary
This adds a DelegatingRunner that will automatically pick irtt for UDP RR
measurements, falling back to netperf UDP_RR if irtt is not available.

Still needed for this to work properly: A way to deal with the data transform
for netperf tests (from roundtrips to RTT) and having irtt check that it can
talk to the server.

Signed-off-by: Toke Høiland-Jørgensen <toke@toke.dk>