Move data transform logic into runner classes
This is a fairly big change where all data transformations (unit
conversions) move into the runner classes and are applied to both raw
values and the aligned data results.

This is motivated by the need to have a single data series be created by
different runners depending on availability of tooling (because of the
irtt tool being used for UDP RTT measurements). This means that runners
can apply knowledge of what they are parsing to the data conversion
process, which makes a lot more sense than the current split. This also
means that all data transform code is moved entirely to within
runners.py.

Previously, data transforms were not applied to the raw_values array.
This was probably a mistake, since it required that data transforms be
applied at plotting time. This commit changes that, but since it is a
change in the data file format, a data file version bump is added. This
also makes it possible to add conversion logic for backwards
compatibility with old data files (to be added in a subsequent commit).