irtt: Don't try to add timestamps to lost packets
Lost packets don't carry a receive timestamp, which is what we use for the
timestamp value. So don't set the 't' key for lost packets (but set a
'send_time' key instead). Also, catch and log all key errors when parsing irtt
output (instead of crashing on a missing key that was supposed to be there).

Signed-off-by: Toke Høiland-Jørgensen <toke@toke.dk>