combiners: Store number of source data points in series meta
When reducing values down to single point, store the number of source data
points in the new 'orig_n' series meta field. This becomes a list of
numbers, corresponding to the number of data points.

The series metadata value is not saved to the results file, but can be
displayed (and is mostly meant for use in the summary formatter).

Signed-off-by: Toke Høiland-Jørgensen <toke@toke.dk>