resultset: Guess timezone offset from data instead of using system time
Old data files did not use UTC timestamps, which meant there was an offset
from the data series values to the ISO timestamps in the metadata. Try to
guess the actual offset instead of using system time, as otherwise old data
files will break if loaded on a machine with a different time zone than the
one it was produced on.

Signed-off-by: Toke Høiland-Jørgensen <toke@toke.dk>