plotters: Don't enforce data series 'units' match when using raw keys
The data pulled out by raw_key plots does not necessarily use the same
units as the data series definition, so this can lead to spurious errors.
For now, just disable the check when raw_key is set, until we figure out a
better way to deal with this.

Fixes #173.

Signed-off-by: Toke Høiland-Jørgensen <toke@toke.dk>