aggregators: Make sure series metadata is actually saved
The call to results.create_series() also clears out the series metadata.
Since this happened after the metadata was saved, the end result was empty
SERIES_META in the data files.

This in turn led to odd results for bandwidth summary statistics, and of
course the absence of all this metadata from the test results, as reported
by Rich Brown.

Fixes #182 (for new data files).

[Backport of ff465203dd7de78d87e0d88717068621b8159c28 from master]

Signed-off-by: Toke Høiland-Jørgensen <toke@toke.dk>