rrul: Make sure to include all ping flows in plots
Use globbing to make sure we also include extra pings added with the
ping_hosts test parameter. This is especially important as those extra
flows are also included in the average, so if they are not plotted, the
average looks off.

Also add 'id' markings to the ping flows so the labels are displayed
properly, and add the type to the extra pings' ID.

Signed-off-by: Toke Høiland-Jørgensen <toke@toke.dk>