runners: When killing subprocesses, try SIGINT before sending SIGTERM
SIGTERM prevents programs from printing end-of-test output, so when killing
processes from kill events, send a SIGINT and give the program half a
second to exit before proceeding with SIGTERM.

Since fping is often killed by the watchdog timer, this should hopefully
restore the series metadata (such as MEAN_VALUE) to runs where this
happens. Which in turn makes ping tests counted in MEAN_VALUE averages and
sums, which they were not previously, as reported in #184.

[ Backport of 67ac0f73966f1a15a19617bcac61099d27b4531b from master ]

Signed-off-by: Toke Høiland-Jørgensen <toke@toke.dk>