gui: Fix plot selection issues when running tests from the GUI
There were a couple of issues with plot selection after running tests from
the GUI, which have to do with state being updated in the settings object
as a side effect of other actions. This change can linger for later test
runs, and look like an explicit plot selection which will produce
a (harmless, but still annoying) error. The issues fixed are:

- The ListTests action will modify the settings object being passed to it.
  Fix this by copying it first.

- The test run dialog in the GUI also keeps a settings object around; make
  sure to reset this at the end of a test run so it doesn't interfere with
  the next test being run.

This fixes #193.

[ Backport of commit a0f2247bfa10b56939d775fe210e98f5aa1f285f from master ]

Signed-off-by: Toke Høiland-Jørgensen <toke@toke.dk>