tests: Support setting UDP packet size for UDP tests
This adds support for a new test parameter 'udp_pktsize' for passing the
desired pktsize to iperf when running the udp_flood* tests. Also add
support for specifying per-stream bandwidths and pktsizes using the
plural-version of the test parameters.

Since this ports the udp_flood_var_up test to use the add_streams() helper,
this also means that the stream_delays test parameter is now also honoured
for UDP streams.

Signed-off-by: Toke Høiland-Jørgensen <toke@toke.dk>