tests: Add rrul_var configurable version of RRUL test
This adds a new version of the RRUL test that is configurable. We now
support per-stream configuration of markings and CC algorithms in some
tests, but not in RRUL test. This adds a new version of RRUL that does: It
uses the bidir_streams test parameter to set the number of streams,
where each 'stream' consists of both an upload a download and an UDP RR
flow.

The rrul_var test supports the existing test parameters to configure
per-stream properties: 'labels', 'markings', 'cc_algos', 'send_sizes' and
'stream_delays'. All three flows will get the same set of markings.

Signed-off-by: Toke Høiland-Jørgensen <toke@toke.dk>