scripts/ss_iterate.sh: Add extra newline after ss output
A bug in certain versions of 'ss' causes its output to miss a newline at
the end. While this has been fixed in upstream iproute2, there are versions
of ss in the wild with this bug. So add an extra newline after the ss
output in ss_iterate.sh; parsing of the output is entirely regex-based, so
an extra newline doesn't hurt.

Fixes #204.

Reported-by: Olivier Tilmans <olivier.tilmans@nokia-bell-labs.com>
Signed-off-by: Toke Høiland-Jørgensen <toke@toke.dk>