settings: Support specifying the same hostname multiple times
Flent has never supported specifying the same hostname multiple times (for
multi-stream tests), because the hostname is also used as a dictionary key.
However, it turns out that it is possible to fix this by adding a suffix to
the hostname if it is specified multiple times, and just stripping out the
suffix before passing the hostname to the runners.

This commit implements that idea, using numbered suffixes formatted as
'//N' (two slashes to avoid the suffix looking like a netmask when applied
to IP addresses). The suffix is transparently added when the same hostname
is specified multiple times, and stripped back out in the runner code.
Thanks for Felix for the idea.

Suggested-by: Felix Fietkau <nbd@nbd.name>
Signed-off-by: Toke Høiland-Jørgensen <toke@toke.dk>