settings: Also support wildcard for --remote-hosts
Also support a special wildcard value (*) for the key in --remote-hosts.
The semantics is that a specific value (or range) for a runner index will
take precedence, and the wildcard value will be used for any runners that
don't have an index specified. So this:

--remote-host=*=host1 --remote-host=1-2=host2

will cause runners with indexes 1 and 2 to use 'host2' as their remote, and
all other runners will use host1.

Suggested-by: Felix Fietkau <nbd@nbd.name>
Signed-off-by: Toke Høiland-Jørgensen <toke@toke.dk>