runners: Don't assume path to binaries are the same on remote hosts
When using ssh to run remote runners, we were just running the binary
detection on the local machine and assuming they are the same on the remote
host. This requires hosts to be close to identical, and is not strictly
necessary; instead, we can just shell out to 'ssh $HOST which' and find the
binary, then run the checks on that remote binary. Implement this for
runners that have a remote host set.

Signed-off-by: Toke Høiland-Jørgensen <toke@toke.dk>