util: Make sure AddHost action is a subclass of argparse._AppendAction
The RC file parsing logic looks for instances of argparse._AppendAction to
determine whether a setting variable is supposed to be a list. When
changing the HOSTS action to its own action class, this check broke leading
to wrong parsing of HOSTS in the RC file. Fix this by just making the new
action a subclass of argparse._AppendAction.

Signed-off-by: Toke Høiland-Jørgensen <toke@toke.dk>