aggregators: Protect against overly long iteration due to invalid timestamps
Some versions of fping will output timestamps from CLOCK_MONOTONIC, which
causes the test duration to appear to be on the order of 50 years. Check
for this condition before we try to aggregate, to avoid looping forever; we
warn if the computed duration is more than 10 times the expected, and abort
iteration entirely at a fixed value of 1 million aggregation steps, which
should hopefully avoid this triggering on any legitimate uses.

Signed-off-by: Toke Høiland-Jørgensen <toke@toke.dk>