formatters: Add 99th percentile data to text-based outputs
The text-based output formats (including the default 'summary' format)
currently include mean and median of each data series (as well as some
other statistics in case of the 'stats' formatter). Include the 99th
percentile as well for easy availability. This is mostly useful for latency
measurements, but just include it as a generic output over all data series.

Signed-off-by: Toke Høiland-Jørgensen <toke@toke.dk>