runners: Make sure to pass -4 or -6 to irtt
The IrttRunner didn't propagate the IP version from the 'settings' object
if it was not set in the test definition, which meant that there would
mostly not be anything passed to irtt, which then got to decide for itself.
Besides being inconsistent with the other runners, it also causes problems
because irtt host lookup has its own idea about address family precedence,
preferring IPv4 when both v4 and v6 are present.

Fix this by always passing the address family argument to irtt. Fixes #241.

Signed-off-by: Toke Høiland-Jørgensen <toke@toke.dk>