Added --test-payload option
This commit adds a feature to give the user the option to specify
a custom file as payload to fill netperf's buffers.
Previously, netperf buffers were filled with randomised payload from
the /dev/urandom device file. But now with the --test-payload option
a custom file can also be sent as payload.
Flent will check if the file is readable before passing it to netperf,
and fail the test run if it isn't.
This can be useful in test scenarios where some data needs to be
sent along with the packets when running the provided tests.

Signed-off-by: Hrishikesh Athalye <hathalye7@hotmail.com>