gui: Disconnect logging display widget when it is destroyed
The GUI logging handler can still get log entries after the GUI widget is
destroyed, which will lead to a crash if it attempts to append entries to
the destroyed widget. So set up a signal handler to clear the widget
reference if the widget is destroyed from the Qt side, and don't emit any
log entries if there's no widget.

Signed-off-by: Toke Høiland-Jørgensen <toke@toke.dk>