runners: Only log timer expiry if the kill event isn't set
When a process with a watchdog set exits before the watchdog timer expires,
it kills the watchdog. However, the watchdog would still log a "Timer
expired" log message, which is a bit confusion. So condition the watchdog
log output on the kill event being unset (which means that the timeout did
indeed happen).

Signed-off-by: Toke Høiland-Jørgensen <toke@toke.dk>