runners: cache netperf probing results per remote host
The NetperfDemoRunner probes the netperf binary to establish that it
supports the output modes we need. This is cached, but only for instances
of the runner that don't have a remote_host set. However, the probing is
actually more expensive on remote hosts, so we should cache that as well;
just reusing the existing caching logic but indexing it on the remote_host
value accomplishes this nicely.

Signed-off-by: Toke Høiland-Jørgensen <toke@toke.dk>