runners: Cache ping binary versions after probing
Like for netperf, the PingRunner does a bunch of probing to figure out
which binary to use for a given run. This is expensive especially when done
on remote hosts, so add caching to the PingRunner (keyed on remote host
name), so the lookup only has to be done once.

Signed-off-by: Toke Høiland-Jørgensen <toke@toke.dk>