util: Add file-backed persistent cache dictionary implementation
This adds an implementation of a "persistent dictionary" that can be
serialised to a JSON file and read back in. The idea is to use this for
caching 'which' lookups and the results of runner feature probing. A
support class for namespacing dictionary lookups into a sub-key of the
parent dictionary is also added, so that a single global cache can be used
from different places under separate keys.

Signed-off-by: Toke Høiland-Jørgensen <toke@toke.dk>