settings: Add new --cache-file argument for instantiating global cache
This adds a new --cache-file argument to the settings object that will
instantiate the file-backed caching dictionary backed by the file argument
specified. The existing 'which' lookup cache is moved into the persistent
cache (under the 'which') key after instantiating. A subsequent commit will
move the runner cache into the same global cache.

Signed-off-by: Toke Høiland-Jørgensen <toke@toke.dk>