runners: Use global persistent cache for storing result of runner probing
This adds a global 'cache' member to the base runner class that will either
point to a regular dictionary, or into the global file-backed cache
dictionary if one has been specified. If using the global cache, the key
will be the class name so every instance of the same runner shares the same
cache. Runner subclasses can also set the 'cls_cache' class member to get a
per-class regular dictionary when no persistent cache is specified;
otherwise, the default will be a fresh dictionary per instance.

Signed-off-by: Toke Høiland-Jørgensen <toke@toke.dk>