util: Better error handling when reading cache file
Improve the error handling when reading the cache file (if enabled): Don't
override the file on parse or IO errors, only if it's non-existent or
empty. On any other error, just log a warning and disable the cache file
entirely. Also try to write the file immediately if it doesn't exist, so we
catch any write errors early and can log them (and disable the cache file).

Signed-off-by: Toke Høiland-Jørgensen <toke@toke.dk>