util: Rework cache to be a singleton object in the util module
Storing the cache instance in the settings object gives all kinds of
problems because the settings object is routinely copied, pickled and
unpickled, leading to inconsistent cache views and the cache file being
constantly overwritten with partial values.

Instead, move the cache object to be a single persistent object initialised
on import of the 'util' object. The cache object is initially imported
without any file backing, and will just function as a dictionary. If the
command-line parameter is specified, the settings module will call the
read_file() method on the cache object, which will cause the cache to
switch to being file-backed.

This also has the benefit that the runner code can just unconditionally use
the cache object; we do this in a property on the base runner object, so
that only runners that actually use the cache will instantiate keys in the
cache dictionary.

We also add locking for updating and writing the cache file, so the cache
object can be used across threads.

Signed-off-by: Toke Høiland-Jørgensen <toke@toke.dk>