gui: Try harder to exit in test_mode
When running the GUI in test_mode from the unittests, it should exit
immediately. Try a bit harder by connecting the exit timer directly to
QApplication::quit() with a QueuedConnection; this should prevent the
signal from being lost if the timer fires before the event loop is started.
Also move the timer connection up to before the main window is shown.

Signed-off-by: Toke Høiland-Jørgensen <toke@toke.dk>