unittests: Create a preforking TestCase instead of using per-method decorator
Instead of having a decorator that does a prefork and is applied to every
test case method, implement a TestCase subclass that does the fork before
calling its parent's run() method. This better encapsulates the test case,
since the setUp() method will also be run in the child process, and it
means we don't have to put decorators everywhere.

With this change, we can move the import of the 'plotters' module into the
setUp() method of the plotting test cases, to make sure initialisation of
matplotlib is performed after forking (so no global state will be leftover
for subsequent runs). This should fix the GUI plotting test cases, when
combined with also turning the test in test_gui into a forking test (to
prevent loading of the GUI to interfere with subsequent GUI-based plotting
because of leftover state from PyQt).

Signed-off-by: Toke Høiland-Jørgensen <toke@toke.dk>