github/workflows: Use xvfb-action instead of roll-your-own
In an attempt to get rid of intermittent errors about invalid $DISPLAY from
matplotlib, switch to using the xvfb-action Github action for running the
tests instead of having a separate step that sets up xvfb. We keep the
custom apt installation of xvfb, though, because we need the extra
components to get the xcb backend to work.

Signed-off-by: Toke Høiland-Jørgensen <toke@toke.dk>