aggregators, runners: Use a multiprocessing Pool for parsing output
Instead of having each runner immediately parse the results in the
supervisor thread, defer the parsing and do a separate pass through the
runners to parse the results from the aggregator. This has the benefit that
we can use a multiprocessing Pool to do the parsing, spreading it over
multiple cores, instead of having all the different Python threads contend
for the GIL while they're all parsing at the same time.

Signed-off-by: Toke Høiland-Jørgensen <toke@toke.dk>