runners: Implement debug() object method that includes class name
There are a lot of places in the runner debug logging where we include the
class name of the runner outputing the message. Let's add a utility
function for that and move all debug logging to use it so the runner name
is added consistently.

Rename the instance variable to 'runner_name' to not interfere with the
'name' property of the Thread object (because that gets sad when accessing
its name after the object has been pickled).

Signed-off-by: Toke Høiland-Jørgensen <toke@toke.dk>