aggregators: Perform a separate pass through runners calling fork()
The ProcessRunner class would fork() as part of its start() method;
however, because a fork inherits its parent's environment, this meant that
subsequent subprocesses would inherit all the threads started before it.
This slows down subsequent forks, which leads to scalability issues.

Prevent this by performing a separate pass through the runners and calling
their fork() method (if it exists).

Signed-off-by: Toke Høiland-Jørgensen <toke@toke.dk>