runners: Simplify timeout handling and killing of children
The logic around killing child processes and in particular around timeouts
was needlessly convoluted. Simplify this by getting rid of the TimerRunner
entirely (no reason to have a whole separate thread just to run a timer) in
favour of just timing the process in the monitor loop. Retain the
kill_event to use for when the main process is interrupted and needs to
terminate. When using it this way we can also get rid of the separate
kill() implementation in the ProcessRunner entirely; setting the event is
enough, the reason it didn't work before was that the aggregator didn't
join() the thread after calling its kill method; fix that as well.

Signed-off-by: Toke Høiland-Jørgensen <toke@toke.dk>