runners: Only read runner output into memory when it is needed
Instead of reading the runner output into memory immediately after the
runner has exited, let's try to defer it until we actually need the output.
We already spool the output to temporary files and read it back from there,
so this is just a matter of keeping the file descriptors open and only
reading the contents right before we need to.

In particular, since we now moved parsing to separate worker processes this
means that unless debug logging is enabled, the main process will not read
the runner output at all unless an error occurs. For this to work, we need
to be able to pickle and unpickle the stdout and stderr members;
fortunately, the multiprocessing stdlib module already includes functions
to make this possible (by sending file descriptors across the socket used
to communicate with child processes), so we just need to use this in the
__getstate__ and __setstate__ methods of the runners.

Signed-off-by: Toke Høiland-Jørgensen <toke@toke.dk>