runners: Turn parsers into stream parsers
Now that we only read the output from stdout and stderr into memory at the
last minute where we need it, the natural next step is to turn the parsers
into stream parsers instead of reading the whole input into memory at once.

This turned out to be fairly simple as most of the parsers were already
very close to being stream parsers (either splitting the output by line or
by chunk delimited by '---'). Both cases are quite simple to turn into real
stream parsers that process each chunk (or line) as they are read. The only
parser that didn't work this way was the IrttRunner which parses JSON
output; but here we can just switch from using json.loads() to json.load()
and have the JSON parser read the data directly from the file.

Signed-off-by: Toke Høiland-Jørgensen <toke@toke.dk>