loggers: Limit the number of lines of tool output on the console
When a parser can't find any valid data it outputs a log message which
will end up including the runner output. However, if there is a lot of
output this can overwhelm the console making the output useless, so let's
limit the number of lines of output in such a message to 100.

Signed-off-by: Toke Høiland-Jørgensen <toke@toke.dk>