packaging/rpm: Fix recommended dependencies
We already had a Recommends on python3-qtpy, but the package is spelled as
python3-QtPy on Fedora. Also, it doesn't actually depend on any backends,
and Fedora ships PySide2, not PyQt5, so change that recommend as well.
Finally, the qt5 backend of matplotlib is also split out into its own
package, so add that along with the others.

Signed-off-by: Toke Høiland-Jørgensen <toke@toke.dk>