aggregators: Limit size of log queue and read it continuously
When parsing data in the multiprocess runner, we were using an
infinite-size queue for log messages from the work threads, and only
reading from it after parsing finished. It seems this can lead to an OOM
condition which shows up as the parsing process hanging. To fix this, limit
the size of the logging queue to 10 items, and run the reading side in a
separate daemon thread so we drain in continuously.

Signed-off-by: Toke Høiland-Jørgensen <toke@toke.dk>