aggregators: Wait for parsing results and process log in parallel
Instead of dumping all log entries into an infinite-size queue when parsing
output in parallel, limit the size of the queue and drain it continuously
from a separate thread. Also wait explicitly from all the AsyncResult
objects returned from the job submissions instead of just waiting for the
pool at the end.

Signed-off-by: Toke Høiland-Jørgensen <toke@toke.dk>