aggregators: Don't use parallel parsing on OSX
It seems the multiprocessing module is a bit flaky on OSX, which leads to
hangs in the output parsing when running on that platform. Since attempts
to find a workaround have thus far failed, let's just disable the parallel
processing entirely on OSX and simply run the parsing sequentially in the
main process.

Signed-off-by: Toke Høiland-Jørgensen <toke@toke.dk>