runners: Try multiple targets when testing fping output format
On OSX, the firewall can be set to "stealth mode", which causes pings
against localhost to fail. We rely on a localhost ping to test the output
format of fping, so in this firewall mode, fping becomes unusable. Fix this
by adding additional ping targets: the selected hostname, and the
Cloudflare public DNS servers. Since fping supports multiple ping targets,
we just supply all of them, and set a relatively short timeout and no
retries to make sure we don't take too long trying to reach any of the
targets.

Fixes #271.

Signed-off-by: Toke Høiland-Jørgensen <toke@toke.dk>