runners: Return valid results from SilentProcessRunner
The SilentProcessRunner is supposed to return an empty resultset, but
it's parse() method just returns None. However, the parsing code expects a
three-argument tuple of dicts and lists, and barfs on the bare None,
leading to a crash in the parsing runner, which leads to Flent hanging
during the run.

Fix this by returning the proper three-element tuple, with the right empty
elements in the dictionary items.

Signed-off-by: Toke Høiland-Jørgensen <toke@toke.dk>