iperf compatibility with iperf 2.0.14a and newer
The iperf binary deprecated the --enhancedreports from version 2.0.14a
and onwards with the --enhanced flag.

Version 2.0.14a of iperf2 prints the -h help message to stderr, while
later versions print it to stdout.

More recent versions of iperf2 add CSV headers to the CSV output and
append timezone data to the timestamps.

Signed-off-by: Frey Alfredsson <freysteinn@freysteinn.com>