runners: Don't pass test-specific -H to netperf for UDP_RR test
The UDP_RR test doesn't support a hostname being specified as a test-specific
parameter (after the -- divider). Make sure we handle this correctly, by only
specifying the global -H parameter on UDP_RR.

This also implies that we can't use a separate control_host value for UDP_RR
tests, so make sure to ignore such values on those tests (but emit a warning
about it). Ignoring and emitting a warning is better than erroring out, because
this means control_host can still be used for the TCP runners in the mixed-flow
tests (such as the RRUL test).

Based on the original report and partial patch by Github user @dmfreemon.
Fixes #286.

Signed-off-by: Toke Høiland-Jørgensen <toke@toke.dk>