runners: Write some debug timing information to stderr after forking
To aid in debugging weird timing issues after forking off a runner process,
write a couple of debug lines to stderr after forking, but before exec'ing, so
they will appear in the debug log files. This should otherwise be harmless as
all of the parsers either ignore the error output completely, or filter out the
lines they are interested in using regexes or prefix matching, and so will just
skip over these debug lines.

Signed-off-by: Toke Høiland-Jørgensen <toke@toke.dk>