plotters: Don't rely on distutils module to check matplotlib version
Python 3.12 deprecated the distutils module, which we use to check the version
of matplotlib and error out if it's too old. Arguably, this is a bit overkill
anyway, so replace it with a simple check of the major version by splitting the
version string on periods.

While we're at it, also reject matplotlib version 2.*, since we already removed
that from CI testing.

Signed-off-by: Toke Høiland-Jørgensen <toke@toke.dk>