Added support for per CPU core monitoring
This commit adds multicore support to the stat_iterate.sh script and to
the CpuStatsRunner. The previous script only read the first entry of the
/proc/stat, which gave the average CPU usage for all cores. This
modified script reads all cpu[0-9]+ records, including the average. It
retains the output format of the original script; however, the
implementation has some minor optimizations, which are not that
significant due to how seldom the script runs.

1. It primarily relies on the awk command for the heavy lifting, making
   it easier to read. Awk handles reading the /proc/stat instead of the
cat command.

2. The previous script used the seq command, which consumes more memory
   when multiple iterations occur. Awk handles this instead with a loop.

I tested the script using gawk with and without the Posix compliance
flag (-P), BusyBox v1.36.1 awk, and mawk 1.3.4.

Signed-off-by: Frey Alfredsson <freysteinn@freysteinn.com>