Added targeted missing plots to the unit tests
Some of the test Flent files lack data for specific plots. This lack of
data occurs either because the file did not include the parameter to
create them or because the feature did not exist when it was created.

The unit test framework ignores all plots that might lack data to ensure
they will not fail when unintentional changes break these tests.
However, this also means it will not catch changes that accidentally
break them.

This commit adds functionality to allow defining which plots should be
omitted by the unit tests on a per-file basis, allowing testing of files
that do include the plots.

Signed-off-by: Frey Alfredsson <freysteinn@freysteinn.com>